+++ src/main/java/bridge/BridgeGame.java
@@ -1,6 +1,9 @@
 package bridge;
 
 import bridge.domain.GameData;
+import bridge.dto.HistoryMapDto;
+import bridge.dto.PlayerLocationDto;
+import bridge.dto.PlayerTryCountDto;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
