+++ src/main/java/bridge/dto/HistoryMapDto.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.dto;
 
 import bridge.domain.GameData;
